Add files via upload
This is the first version of the data analysis, the main work is done and please use the crdeit_risk_features.csv to train your models. Further work will be done to clean it up more and make it more readable and presentable. But the data is ready for use